refactor(ros-z): share booster walking conversion
