+++ crates/bevyhavior_simulator/src/config.rs
@@ -1,8 +1,9 @@
 use std::{net::SocketAddr, time::Duration};
 
+use booster::walking::WalkingParameters;
 use color_eyre::{Result, eyre::Context};
 use serde::Deserialize;
-use types::parameters::{BehaviorParameters, RLWalkingParameters};
+use types::parameters::BehaviorParameters;
 
 pub const DEFAULT_TICK_DURATION: Duration = Duration::from_millis(10);
 
@@ -74,15 +75,14 @@ struct BoosterWalkingParameters {
     deceleration_distance: f32,
 }
 
-pub fn default_walking_parameters() -> Result<RLWalkingParameters> {
+pub fn default_walking_parameters() -> Result<WalkingParameters> {
     let file: BoosterInterfaceParametersFile = json5::from_str(include_str!(
         "../../../etc/parameters/ros_z/base/booster_interface.json5"
     ))
     .wrap_err("failed to parse walking parameters")?;
-    Ok(RLWalkingParameters {
+    Ok(WalkingParameters {
         hybrid_align_distance: file.walking.hybrid_align_distance,
         max_alignment_rate: file.walking.max_alignment_rate,
         deceleration_distance: file.walking.deceleration_distance,
-        ..Default::default()
     })
 }
