+++ crates/booster/Cargo.toml
@@ -15,6 +15,7 @@ pyo3 = { workspace = true, optional = true }
 ros-z = { workspace = true }
 ros2 = { workspace = true }
 serde = { workspace = true }
+types = { workspace = true }
 
 [features]
-pyo3 = ["dep:pyo3"]
+pyo3 = ["dep:pyo3", "types/pyo3"]
