+++ crates/types/Cargo.toml
@@ -7,7 +7,6 @@ homepage.workspace = true
 
 [dependencies]
 approx = { workspace = true }
-booster = { workspace = true }
 color-eyre = { workspace = true }
 coordinate_systems = { workspace = true }
 enum-iterator = { workspace = true }
@@ -20,6 +19,7 @@ nalgebra = { workspace = true }
 num-traits = { workspace = true }
 ordered-float = { workspace = true }
 path_serde = { workspace = true }
+pyo3 = { workspace = true, optional = true }
 ros-z = { workspace = true }
 ros2 = { workspace = true }
 serde = { workspace = true }
@@ -27,3 +27,6 @@ splines = { workspace = true }
 
 [dev-dependencies]
 ros-z-schema = { workspace = true }
+
+[features]
+pyo3 = ["dep:pyo3"]
