+++ crates/types/src/lib.rs
@@ -10,6 +10,7 @@ pub mod color;
 pub mod condition_input;
 pub mod cycle_time;
 pub mod detected_feet;
+pub mod fall_down_state;
 pub mod fall_state;
 pub mod field_border;
 pub mod field_color;
